Refactored and renamed the update material system
